# 8343541: C1: Plain memory accesses are emitted with membars with +AlwaysAtomicAccesses