@pengxiaolong 2024-11-18
> It is a bit sad that we call `volatile_field_*` when we really mean atomic. But renaming that would lead to much deeper changes. So I am okay with doing just this, thanks.

I'll fix it, it is not too deep, and they are only used by barrierSetC1.cpp